+++ iam.tf
@@ -22,12 +22,11 @@ data "aws_iam_policy_document" "lambda_basic" {
     effect = "Allow"
 
     actions = [
-      "logs:CreateLogGroup",
       "logs:CreateLogStream",
       "logs:PutLogEvents",
     ]
 
-    resources = ["arn:aws:logs:*:*:*"]
+    resources = [aws_cloudwatch_log_group.lambda[0].arn]
   }
 }
 
@@ -52,6 +51,8 @@ resource "aws_iam_role" "lambda" {
 
   name_prefix        = "lambda"
   assume_role_policy = data.aws_iam_policy_document.assume_role[0].json
+
+  tags = merge(var.tags, var.iam_role_tags)
 }
 
 resource "aws_iam_role_policy" "lambda" {
@@ -68,4 +69,3 @@ resource "aws_iam_role_policy" "lambda" {
     0,
   )
 }
-