+++ main.tf
@@ -8,6 +8,8 @@ resource "aws_sns_topic" "this" {
   count = var.create_sns_topic && var.create ? 1 : 0
 
   name = var.sns_topic_name
+
+  tags = merge(var.tags, var.sns_topic_tags)
 }
 
 locals {
@@ -19,6 +21,16 @@ locals {
     ),
     0,
   )
+}
+
+resource "aws_cloudwatch_log_group" "lambda" {
+  count = var.create ? 1 : 0
+
+  name              = "/aws/lambda/${var.lambda_function_name}"
+  retention_in_days = var.cloudwatch_log_group_retention_in_days
+  kms_key_id        = var.cloudwatch_log_group_kms_key_id
+
+  tags = merge(var.tags, var.cloudwatch_log_group_tags)
 }
 
 resource "aws_sns_topic_subscription" "sns_notify_slack" {
@@ -66,12 +78,13 @@ resource "aws_lambda_function" "notify_slack" {
 
   function_name = var.lambda_function_name
 
-  role             = aws_iam_role.lambda[0].arn
-  handler          = "notify_slack.lambda_handler"
-  source_code_hash = data.archive_file.notify_slack[0].output_base64sha256
-  runtime          = "python3.6"
-  timeout          = 30
-  kms_key_arn      = var.kms_key_arn
+  role                           = aws_iam_role.lambda[0].arn
+  handler                        = "notify_slack.lambda_handler"
+  source_code_hash               = data.archive_file.notify_slack[0].output_base64sha256
+  runtime                        = "python3.7"
+  timeout                        = 30
+  kms_key_arn                    = var.kms_key_arn
+  reserved_concurrent_executions = var.reserved_concurrent_executions
 
   environment {
     variables = {
@@ -82,11 +95,14 @@ resource "aws_lambda_function" "notify_slack" {
     }
   }
 
+  tags = merge(var.tags, var.lambda_function_tags)
+
   lifecycle {
     ignore_changes = [
       filename,
       last_modified,
     ]
   }
-}
 
+  depends_on = ["aws_cloudwatch_log_group.lambda"]
+}