+++ CHANGELOG.md
@@ -3,9 +3,31 @@
 
 
 
+<a name="v2.3.0"></a>
+## [v2.3.0] - 2019-11-08
+
+- Create AWS Cloudwatch log group and give explicit access to it ([#40](https://github.com/terraform-aws-modules/terraform-aws-notify-slack/issues/40))
+
+
+<a name="v2.2.0"></a>
+## [v2.2.0] - 2019-11-08
+
+- Updated CHANGELOG
+- Added support for reserved_concurrent_executions
+
+
+<a name="v2.1.0"></a>
+## [v2.1.0] - 2019-11-08
+
+- Updated CHANGELOG
+- Updated docs, python3.7
+- Add support for resource tagging ([#45](https://github.com/terraform-aws-modules/terraform-aws-notify-slack/issues/45))
+
+
 <a name="v2.0.0"></a>
 ## [v2.0.0] - 2019-06-12
 
+- Updated CHANGELOG
 - Upgraded module to support Terraform 0.12 ([#36](https://github.com/terraform-aws-modules/terraform-aws-notify-slack/issues/36))
 
 
@@ -122,7 +144,10 @@
 - Initial commit
 
 
-[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v2.3.0...HEAD
+[v2.3.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v2.2.0...v2.3.0
+[v2.2.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v2.1.0...v2.2.0
+[v2.1.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v2.0.0...v2.1.0
 [v2.0.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v1.13.0...v2.0.0
 [v1.13.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v1.12.0...v1.13.0
 [v1.12.0]: https://github.com/terraform-aws-modules/terraform-aws-notify-slack/compare/v1.11.0...v1.12.0